feat: editCategoryDialog Server connection

+++ app/src/main/java/kr/dori/android/own_cast/forApiData/AuthResponse.kt
@@ -57,7 +57,7 @@ data class PostCastForResponse(
 data class GetAllPlaylist(
     val name: String,
     val imagePath: String,
-    val playlistId: Int,
+    val playlistId: Long,
     val totalCast: Int
 )
